Refactored java examples platform->deploy
