+++ vertx-examples/src/main/java/http/ServerExample.java
@@ -18,7 +18,7 @@ package http;
 
 import org.vertx.java.core.Handler;
 import org.vertx.java.core.http.HttpServerRequest;
-import org.vertx.java.platform.Verticle;
+import org.vertx.java.deploy.Verticle;
 
 public class ServerExample extends Verticle {
