+++ vertx-examples/src/main/java/httpperf/PerfServer.java
@@ -21,7 +21,7 @@ import org.vertx.java.core.AsyncResultHandler;
 import org.vertx.java.core.Handler;
 import org.vertx.java.core.buffer.Buffer;
 import org.vertx.java.core.http.HttpServerRequest;
-import org.vertx.java.platform.Verticle;
+import org.vertx.java.deploy.Verticle;
 
 public class PerfServer extends Verticle {
