+++ vertx-examples/src/main/java/https/ClientExample.java
@@ -19,7 +19,7 @@ package https;
 import org.vertx.java.core.Handler;
 import org.vertx.java.core.buffer.Buffer;
 import org.vertx.java.core.http.HttpClientResponse;
-import org.vertx.java.platform.Verticle;
+import org.vertx.java.deploy.Verticle;
 
 public class ClientExample extends Verticle {
