+++ vertx-examples/src/main/java/perf/RateCounter.java
@@ -18,7 +18,7 @@ package perf;
 
 import org.vertx.java.core.Handler;
 import org.vertx.java.core.eventbus.Message;
-import org.vertx.java.platform.Verticle;
+import org.vertx.java.deploy.Verticle;
 
 public class RateCounter extends Verticle implements Handler<Message<Integer>> {
