+++ vertx-examples/src/main/java/sendfile/SendFileExample.java
@@ -18,7 +18,7 @@ package sendfile;
 
 import org.vertx.java.core.Handler;
 import org.vertx.java.core.http.HttpServerRequest;
-import org.vertx.java.platform.Verticle;
+import org.vertx.java.deploy.Verticle;
 
 public class SendFileExample extends Verticle {
