+++ vertx-examples/src/main/java/sourcemod/mods/test-mod-v1.0/org/foo/MyMain.java
@@ -3,7 +3,7 @@ package org.foo;
 import com.acme.OtherClass;
 import org.vertx.java.core.Handler;
 import org.vertx.java.core.http.HttpServerRequest;
-import org.vertx.java.platform.Verticle;
+import org.vertx.java.deploy.Verticle;
 
 public class MyMain extends Verticle {
