+++ vertx-examples/src/main/java/ssl/SSLClient.java
@@ -19,7 +19,7 @@ package ssl;
 import org.vertx.java.core.Handler;
 import org.vertx.java.core.buffer.Buffer;
 import org.vertx.java.core.net.NetSocket;
-import org.vertx.java.platform.Verticle;
+import org.vertx.java.deploy.Verticle;
 
 public class SSLClient extends Verticle {
