+++ vertx-examples/src/main/java/wsperf/ConnectClient.java
@@ -19,7 +19,7 @@ package wsperf;
 import org.vertx.java.core.Handler;
 import org.vertx.java.core.http.HttpClient;
 import org.vertx.java.core.http.WebSocket;
-import org.vertx.java.platform.Verticle;
+import org.vertx.java.deploy.Verticle;
 
 public class ConnectClient extends Verticle {
