+++ vertx-examples/src/main/java/wsperf/PerfServer.java
@@ -19,7 +19,7 @@ package wsperf;
 import org.vertx.java.core.Handler;
 import org.vertx.java.core.http.ServerWebSocket;
 import org.vertx.java.core.streams.Pump;
-import org.vertx.java.platform.Verticle;
+import org.vertx.java.deploy.Verticle;
 
 /**
  * @author <a href="http://tfox.org">Tim Fox</a>
